fix: simplify task queue progress, which probably fixes a few bugs

+++ src/main/java/net/technicpack/launchercore/mirror/download/Download.java
@@ -58,7 +58,7 @@ public class Download implements Runnable {
 
     public float getProgress() {
         if (size > 0) {
-            return ((float) downloaded / size) * 100;
+            return (float) ((double) downloaded / size * 100);
         } else {
             return 0;
         }
